# apps/api/openapi/spec.yaml
openapi: 3.0.0
info:
  title: lomi.africa Payments API
  version: 1.0.0
  description: API of lomi. payment orchestration platform
servers:
  - url: https://api.lomi.africa/v1
    
# API Schema to Database Table Mapping:
# - Merchant (API) -> merchants (DB)
# - Product (API) -> merchant_products (DB)
# - SubscriptionPlan (API) -> subscription_plans (DB)
# - Subscription (API) -> merchant_subscriptions (DB)
# - Transaction (API) -> transactions (DB)
# - Customer (API) -> customers (DB)
# - PaymentLink (API) -> payment_links (DB)
# - Refund (API) -> refunds (DB)
# - Webhook (API) -> webhooks (DB)
# - CheckoutSession (API) -> checkout_sessions (DB)
# - MerchantAccount (API) -> merchant_accounts (DB)
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    Provider:
      type: object
      properties:
        code:
          $ref: '#/components/schemas/ProviderCode'
        name:
          type: string
        description:
          type: string
        payment_methods:
          type: array
          items:
            $ref: '#/components/schemas/PaymentMethodCode'
        is_connected:
          type: boolean
    CurrencyCode:
      type: string
      enum:
        - XOF
        - USD
        - EUR
        - GHS
        - NGN
        - KES
        - MRO
...
    ProviderCode:
      type: string
      enum:
        - ORANGE
        - WAVE
        - ECOBANK
        - MTN
        - NOWPAYMENTS
        - APPLE
        - GOOGLE
        - MOOV
        - AIRTEL
        - MPESA
        - OPAY
        - PAYPAL
        - OZOW
        - OTHER
    PaymentMethodCode:
      type: string
      enum:
        - CARDS
        - MOBILE_MONEY
        - E_WALLET
        - APPLE_PAY
        - GOOGLE_PAY
        - USSD
        - QR_CODE
        - BANK_TRANSFER
        - CRYPTO
        - PAYPAL
        - OTHER
...
security:
  - ApiKeyAuth: []
paths:
...
  /providers:
    get:
      summary: List available payment providers
      operationId: providers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Provider'
        '401':
          $ref: '#/components/responses/UnauthorizedError' 
...
  /merchants/{merchant_id}/mrr:
    get:
      summary: Get merchant monthly recurring revenue
      description: Retrieve current monthly recurring revenue
      operationId: getMerchantMrr
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Monthly recurring revenue retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant_id:
                    type: string
                    format: uuid
                  mrr:
                    type: number
                    description: Monthly Recurring Revenue
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /merchants/{merchant_id}/arr:
    get:
      summary: Get merchant annual recurring revenue
      description: Retrieve current annual recurring revenue
      operationId: getMerchantArr
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Annual recurring revenue retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant_id:
                    type: string
                    format: uuid
                  arr:
                    type: number
                    description: Annual Recurring Revenue
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /merchants/{merchant_id}/balance:
    get:
      summary: Get merchant account balance
      description: Retrieve current account balance for a merchant
      operationId: getMerchantBalance
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: currency_code
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CurrencyCode'
          description: Filter balance by currency
      responses:
        '200':
          description: Account balance retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    merchant_id:
                      type: string
                      format: uuid
                    currency_code:
                      $ref: '#/components/schemas/CurrencyCode'
                    balance:
                      type: number
                    created_at:
                      type: string
                      format: date-time
                    updated_at:
                      type: string
                      format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...